Give command some structure with cobra
This gives the CLI the suggestion of structure. The intended shape is
something like:

    flux-whatif merge --url ... --branch ... --new-ref ...

That is: "what if the HEAD of branch <branch> in the git repo at <url>
pointed to <new-ref>" -- the general case of "what if I merged this
PR".

I find cobra helps organise things when you get into subcommands, so
I've used that.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>