Factor out the initial scenario
This factors the question of the initial scenario out of the
processing loop, into its own abstraction.

The aim is that the simulation can be generic with respect to the
scenario. The first element of this is finding the sources that are
affected, so that becomes a method of the scenario abstraction. There
is a bit more to do here, since constructing the artifacts also
depends on the scenario.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>