Return the diffs as structs, then print
To be able to analyse whether changes will provoke further changes, we
need to know if any of the changed objects are Flux objects.

This commit alters the copied build/ package so that Diff returns all
the differences as struct values, rather than printing them out as it
goes. That way, I can print them, _and_ examine whether there is
another round necessary.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>