Rely on mutating GitRepository objects
Instead of passing a new ref through all the funcs handling git
clones, just mutate the GitRepository object so that the code does the
right thing.

This is a bit neater, and means fewer alterations to the copied
controller code; but it also means they will all work when we have
(pretend) patched versions from the live system, rather than sources
altered according to the scenario.

This also factors out the code common to filtering down the git repos
from the scenario.

Signed-off-by: Michael Bridgen <mikeb@squaremobius.net>